+++ src/main/java/ControlledMoney/api/domain/lucro/dtos/LucroInputDTO.java
@@ -16,7 +16,7 @@ public record LucroInputDTO(
     String motivo,
 
     @NotNull
-    Long valor,
+    Double valor,
     
     @NotBlank
     Boolean recebido,
